requirements.txt
faster-whisper
piper-voice
sounddevice
scipy
numpy
pvporcupine
pyaudio
aiohttp
